backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic-settings==2.1.0

# Database
# sqlalchemy==2.0.23  # Replaced with Cosmos DB
# pymysql==1.1.0  # Replaced with Cosmos DB
# alembic==1.12.1  # Not needed for NoSQL
azure-cosmos==4.5.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Stock Data
# yfinance==0.2.31  # Replaced with Alpha Vantage
requests==2.31.0

# Configuration
pydantic==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
ruff==0.1.6
black==23.11.0
mypy==1.7.0
